requirements.txt
customtkinter>=5.2.0
paramiko>=3.4.0
pystray>=0.19.0
cryptography>=42.0.0
Pillow>=10.0.0
pywin32>=306; sys_platform == 'win32'